move store.js from mods/ to manager/
